client/requirements.txt
opencv-python
PyCryptodome
translate